+++ include/GOptimizerPars.hpp
@@ -1,7 +1,7 @@
 /***************************************************************************
  *         GOptimizerPars.hpp - Optimizer parameter container class        *
  * ----------------------------------------------------------------------- *
- *  copyright (C) 2009-2014 by Juergen Knoedlseder                         *
+ *  copyright (C) 2009-2015 by Juergen Knoedlseder                         *
  * ----------------------------------------------------------------------- *
  *                                                                         *
  *  This program is free software: you can redistribute it and/or modify   *
@@ -152,7 +152,7 @@ const GOptimizerPar* GOptimizerPars::operator[](const int& index) const
 inline
 int GOptimizerPars::size(void) const
 {
-    return (m_pars.size());
+    return (int)m_pars.size();
 }
